+++ datawinners/project/web_questionnaire_form_creator.py
@@ -188,7 +188,7 @@ class WebQuestionnaireFormCreater(object):
         return telephone_number_field
 
     def _create_integer_field(self, field, language):
-        integer_field = django.forms.fields.IntegerField(label=field.label[language], required=field.is_required(),
+        integer_field = django.forms.fields.FloatField(label=field.label[language], required=field.is_required(),
             error_messages={'invalid': _('Enter a valid integer')})
         integer_field.widget.attrs["watermark"] = get_integer_field_constraint_text(field)
         integer_field.widget.attrs['style'] = 'padding-top: 7px;'
